Extract run CLI command
